[refactor] 토큰 DB -> Redis 이관

+++ src/main/java/com/hhconcert/server/application/dto/TokenResult.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record TokenResult(
         String tokenId,
         String userId,
-        Integer priority,
+        Long priority,
         TokenStatus status,
         LocalDateTime tokenIssuedAt,
         LocalDateTime activeAt,
